requirements.txt
discord.py==1.7.3
discord-py-slash-command==1.1.0
python-dotenv
requests